Update int types to Apple spec
git-svn-id: svn://svn.savannah.nongnu.org/gap/trunk@3593 d25b4930-3292-e311-b074-0010c6de5da8

+++ system-apps/Addresses/Frameworks/Addresses/ADGroup.h
@@ -30,8 +30,8 @@
 - (NSString*) distributionIdentifierForProperty: (NSString*) property
 					 person: (ADPerson*) person;
 
-+ (int) addPropertiesAndTypes: (NSDictionary*) properties;
-+ (int) removeProperties: (NSArray*) properties;
++ (NSInteger) addPropertiesAndTypes: (NSDictionary*) properties;
++ (NSInteger) removeProperties: (NSArray*) properties;
 + (NSArray*) properties;
 + (ADPropertyType) typeOfProperty: (NSString*) property;
